% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{add_website}
\alias{add_website}
\title{Adds website to Search Console}
\usage{
add_website(siteUrl)
}
\arguments{
\item{siteURL}{The URL of the website to add.}
}
\value{
TRUE if successful, raises an error if not.
}
\description{
Adds website to Search Console
}
\seealso{
Other search console website functions: \code{\link{delete_website}};
  \code{\link{list_websites}}
}

