% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_urlshort.R
\name{analytics_url}
\alias{analytics_url}
\title{Get analyitcs of a url that has used goo.gl}
\usage{
analytics_url(shortUrl, timespan = c("allTime", "month", "week", "day",
  "twoHours"))
}
\arguments{
\item{shortUrl}{Url that was shortened with goo.gl}

\item{timespan}{The time period for the analytics data}
}
\value{
a dataframe of the goo.gl Url analytics
}
\description{
Get analyitcs of a url that has used goo.gl
}

