% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{applyDataParseFunction}
\alias{applyDataParseFunction}
\title{Apply parsing function if a good response}
\usage{
applyDataParseFunction(function_entry, batch_content)
}
\description{
Apply parsing function if a good response
}

