% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{authReturnCode}
\alias{authReturnCode}
\title{Returns the authentication parameter "code" in redirected URLs}
\usage{
authReturnCode(session, securityCode = getOption("googleAuthR.securitycode"))
}
\arguments{
\item{session}{A shiny session object}

\item{securityCode}{A random string to check the auth comes form the same origin.}
}
\value{
The Google auth token in the code URL parameter.
}
\description{
Checks the URL of the Shiny app to get the state and code URL parameters.
}
\seealso{
Other shiny auth functions: \code{\link{createCode}},
  \code{\link{gar_shiny_getAuthUrl}},
  \code{\link{gar_shiny_getToken}},
  \code{\link{gar_shiny_getUrl}},
  \code{\link{loginOutput}},
  \code{\link{reactiveAccessToken}},
  \code{\link{renderLogin}},
  \code{\link{revokeEventObserver}},
  \code{\link{with_shiny}}
}
\keyword{internal}
