% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{check.Url}
\alias{check.Url}
\title{Checks Urls are in right format for API request}
\usage{
check.Url(url, checkProtocol = TRUE)
}
\arguments{
\item{url}{The URL to check for use in API request}

\item{checkProtocol}{Check if URL starts with 'http'}

\item{...}{Passed to URLencode()}
}
\value{
Modified url if successful, raises an error if not.
}
\description{
Checks Urls are in right format for API request
}
\keyword{internal}
