% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkGoogleAPIError}
\alias{checkGoogleAPIError}
\title{Get Google API errors}
\usage{
checkGoogleAPIError(req)
}
\arguments{
\item{req}{a httr request}
}
\description{
Get Google API errors
}
\keyword{internal}
