% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_generator.R
\name{checkTokenAPI}
\alias{checkTokenAPI}
\title{Check API data token}
\usage{
checkTokenAPI(shiny_access_token = NULL, verbose = F)
}
\arguments{
\item{shiny_access_token}{auth token}
}
\value{
boolean if it works.
}
\description{
Check API data token
}
\keyword{internal}

