% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{createCode}
\alias{createCode}
\title{Creates a random character code}
\usage{
createCode(seed = NULL, num = 20)
}
\arguments{
\item{seed}{random seed.}

\item{num}{number of characters the code should be.}
}
\value{
a string of random digits and letters.
}
\description{
Creates a random character code
}
\seealso{
Other shiny auth functions: \code{\link{authReturnCode}},
  \code{\link{gar_shiny_getAuthUrl}},
  \code{\link{gar_shiny_getToken}},
  \code{\link{gar_shiny_getUrl}},
  \code{\link{loginOutput}},
  \code{\link{reactiveAccessToken}},
  \code{\link{renderLogin}},
  \code{\link{revokeEventObserver}},
  \code{\link{with_shiny}}
}
\keyword{internal}
