% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{doBatchRequest}
\alias{doBatchRequest}
\title{doBatchRequest}
\usage{
doBatchRequest(req_url, shiny_access_token, http_header, the_body)
}
\description{
Returns whats necessary for a batch request
}

