% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_generator.R
\name{doHttrRequest}
\alias{doHttrRequest}
\title{Get URL content based on if its Shiny or local}
\usage{
doHttrRequest(url, shiny_access_token = NULL, request_type = "GET",
  the_body = NULL)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{request_type}{the type of httr request function: GET, POST, PUT, DELETE etc.}

\item{the_body}{body of POST request}

\item{session}{a shiny session object if in a shiny app.}

\item{params}{A named character vector of other parameters to add to request.}
}
\description{
This changes the auth type depending on if its local or on Shiny
}
\details{
Example of params: c(param1="foo", param2="bar")
}
\keyword{internal}

