% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_urlshort.R
\name{expand_url}
\alias{expand_url}
\title{Expands a url that has used goo.gl}
\usage{
expand_url(shortUrl)
}
\arguments{
\item{shortUrl}{Url that was shortened with goo.gl}
}
\value{
a string of the expanded URL
}
\description{
Expands a url that has used goo.gl
}

