% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_generator.R
\name{gar_api_generator}
\alias{gar_api_generator}
\title{googleAuthR data fetch function generator}
\usage{
gar_api_generator(baseURI, http_header = c("GET", "POST", "PUT", "DELETE",
  "PATCH"), path_args = NULL, pars_args = NULL,
  data_parse_function = NULL)
}
\arguments{
\item{baseURI}{The stem of the API call.}

\item{http_header}{Type of http request.}

\item{path_args}{A named list with name=folder in request URI, value=the function variable.}

\item{pars_args}{A named list with name=parameter in request URI, value=the function variable.}

\item{data_parse_function}{A function that takes a request response, parses it and returns the data you need.

\strong{path_args} and \strong{pars_args} add default values to the baseURI.
You don't need to supply access_token for OAuth2 requests in pars_args,
  this is dealt with in gar_auth()}
}
\value{
A function that can fetch the Google API data you specify
}
\description{
This function generates other functions for use with Google APIs
}

