% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_results.R
\name{gar_api_page}
\alias{gar_api_page}
\title{Takes a generated API function and lets you page through results}
\usage{
gar_api_page(f, next_f = function(x) x$nextPageToken,
  page_arg = "start-index")
}
\arguments{
\item{f}{a function created by gar_api_generator}

\item{next_f}{A function that will let you get the nextPageToken from objects returned by f}

\item{page_arg}{In f, you need to set a pars_argument that will change for each page.}
}
\value{
A list of the API responses
}
\description{
A helper function to help with the common task of paging through large API results.
}
\details{
The next_token_f needs to extract the nextPageToken from the API response after the data_parse_function has been applied.
 e.g. if response is x and next_token_f = function(x) x$npt then the nextPageToken should be available at x$npt after parsing.  It should return NULL if no (more) paging is necessary.
}
