% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_auth_js}
\alias{gar_auth_js}
\title{gar_auth_js UI}
\usage{
gar_auth_js(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{id}{Shiny id}
}
\value{
Shiny UI

A httr reactive OAuth2.0 token
}
\description{
Shiny Module for use with \link{gar_auth_js}
}

