% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_auth_jsUI}
\alias{gar_auth_jsUI}
\title{gar_auth_js UI}
\usage{
gar_auth_jsUI(id, class = "btn btn-success")
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{gar_auth_js}
}

