% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{gar_auth_server}
\alias{gar_auth_server}
\title{gar_auth_server}
\usage{
gar_auth_server(session)
}
\arguments{
\item{session}{Shiny session argument}
}
\description{
If using googleAuth_ui, put this at the top of your server.R function
}
