% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\name{gar_auth_service}
\alias{gar_auth_service}
\title{JSON service account authentication}
\usage{
gar_auth_service(json_file)
}
\arguments{
\item{json_file}{the JSON file downloaded from Google Developer Console}
}
\value{
(Invisible) Sets authentication token
}
\description{
As well as OAuth2 authentication, you can authenticate without user interaction
  via Service accounts.  This involves downloading a secret JSON key with the authentication
  details.
  
  To use, go to your Project in the \href{https://console.developers.google.com/apis/credentials/serviceaccountkey}{Google Developement Console} and select JSON Key type.  Save the file 
  to your computer and call it via supplying the file path to the \code{json_file} parameter.
  
  Navigate to it via: 
    Google Dev Console > Credentials > New credentials > Service account Key > 
       Select service account > Key type = JSON
}
\examples{

\dontrun{
library(googleAuthR)

## secrets is the JSON file downloaded from theGoogle Dev Console
##   Credentials > New credentials > Service account Key > 
##     Select service account > Key type = JSON
service_token <- gar_auth_service(secrets)

analytics_url <- function(shortUrl, 
                       timespan = c("allTime", "month", "week","day","twoHours")){
       timespan <- match.arg(timespan)
       
       f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url",
                      "GET",
                      pars_args = list(shortUrl = "shortUrl",
                                       projection = "FULL"),
                                      data_parse_function = function(x) { 
                                      a <- x$analytics 
                                      return(a[timespan][[1]])
                   })
                   
       f(pars_arguments = list(shortUrl = shortUrl))
       }
 
 analytics_url("https\%3A\%2F\%2Fgoo.gl\%2FcFVQbk")
}


}
\seealso{
https://developers.google.com/identity/protocols/OAuth2ServiceAccount

\code{\link{https://developers.google.com/identity/protocols/OAuth2ServiceAccount}}
}

