% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{gar_batch}
\alias{gar_batch}
\title{Turn a vector of gar_fetch_functions into batch functions}
\usage{
gar_batch(function_list)
}
\description{
Turn a vector of gar_fetch_functions into batch functions
}

