% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_cache.R
\name{gar_cache_list}
\alias{gar_cache_list}
\title{See which functions have mock cache}
\usage{
gar_cache_list(cache_dir = gar_cache_get_loc())
}
\description{
When testing API calls, see which functions have API caches 
  activated when \code{options(googleAuthR.mock_test = TRUE)}
}
