% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoise.R
\name{gar_cache_set_loc}
\alias{gar_cache_set_loc}
\alias{gar_cache_get_loc}
\alias{gar_cache_empty}
\title{Set cache location}
\usage{
gar_cache_set_loc(cache)

gar_cache_get_loc()

gar_cache_empty()
}
\arguments{
\item{cache}{The directory to save cache to, or \code{"memory"} to save to RAM}
}
\description{
These functions let you set the cache behaviour for your API calls
}
