% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_discovery.R
\name{gar_create_api_objects}
\alias{gar_create_api_objects}
\title{Create the API objects from the Discovery API}
\usage{
gar_create_api_objects(filename = "./inst/api_objects.R", api_json)
}
\arguments{
\item{filename}{File to write the objects to}

\item{api_json}{The json from \link{gar_discovery_api}}
}
\value{
TRUE if successful, side-effect creating filename
}
\description{
Create the API objects from the Discovery API
}

