% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_discovery.R
\name{gar_create_package}
\alias{gar_create_package}
\title{Create a Google API package}
\usage{
gar_create_package(api_json, directory, rstudio = TRUE)
}
\arguments{
\item{api_json}{json from \link{gar_discovery_api}}

\item{directory}{Where to build the package}

\item{rstudio}{Passed to \link[devtools]{create}, creates RStudio project file.

Uses \link[devtools]{create} to create a package structure then 
  \link{gar_create_api_skeleton} and \link{gar_create_api_objects} to create 
  starting files for a Google API package.}
}
\description{
Create a Google API package
}
\seealso{
\href{https://developers.google.com/discovery/v1/reference/apis/list}

Other Google Discovery API functions: \code{\link{gar_create_api_objects}},
  \code{\link{gar_create_api_skeleton}},
  \code{\link{gar_discovery_apis_list}},
  \code{\link{gar_discovery_api}}
}

