% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_gargle.R
\name{gar_deauth}
\alias{gar_deauth}
\title{Suspend authorization}
\usage{
gar_deauth()
}
\description{
Put googleAuthR into a de-authorized state. Instead of sending a token,
googleAuthR will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
googleAuthR in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via [gar_auth_configure()] and retrieve that key via
[gar_api_key()].
In the absence of a user-configured key, a built-in default key is used.
}
\examples{
\dontrun{
gar_deauth()
}
}
\seealso{
Other auth functions: 
\code{\link{gar_auth_configure}()}
}
\concept{auth functions}
