% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugging.R
\name{gar_debug_parsing}
\alias{gar_debug_parsing}
\title{Read the diagnostic object returned on API parse errors.}
\usage{
gar_debug_parsing(filename = "gar_parse_error.rds")
}
\arguments{
\item{filename}{The file created from API errors, usually called \code{gar_parse_error.rds}}
}
\description{
Read the diagnostic object returned on API parse errors.
}
\details{
When googleAuthR API parsing fails, it will write a file called gar_parse_error.rds to the directory.  Feed that file into this function to help diagnose the problem.
}
