% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_discovery.R
\name{gar_discovery_api}
\alias{gar_discovery_api}
\title{Get meta data details for specified Google API}
\usage{
gar_discovery_api(api, version)
}
\arguments{
\item{api}{The API to fetch}

\item{version}{The API version to fetch}
}
\value{
Details of the API
}
\description{
Doesn't require authentication
}
\seealso{
\href{https://developers.google.com/discovery/v1/reference/apis/getRest}
}

