% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{gar_discovery_apis_list}
\alias{gar_discovery_apis_list}
\title{Get a list of Google API libraries}
\usage{
gar_discovery_apis_list()
}
\value{
List of Google APIs and their resources
}
\description{
Doesn't require authentication
}
\seealso{
\href{https://developers.google.com/discovery/v1/reference/apis/list}
}

