% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gce.R
\name{gar_gce_auth_email}
\alias{gar_gce_auth_email}
\title{Get the service email via GCE metadata}
\usage{
gar_gce_auth_email(service_account = "default")
}
\arguments{
\item{service_account}{Specify a different service account from the \code{default}

Useful if you don't know the default email and need it for other uses}
}
\value{
the email address character string
}
\description{
Get the service email via GCE metadata
}
\seealso{
\link{gar_gce_auth}
}
