% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_gargle.R
\name{gar_has_token}
\alias{gar_has_token}
\title{Is there a token on hand?}
\usage{
gar_has_token()
}
\value{
Logical.
}
\description{
Reports whether googleAuthR has stored a token, ready for use in downstream
requests.
}
\examples{
gar_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{gar_token}()}
}
\concept{low-level API functions}
