% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_js_getToken}
\alias{gar_js_getToken}
\title{Create a httr token from a js token}
\usage{
gar_js_getToken(token, client.id = getOption("googleAuthR.webapp.client_id"),
  client.secret = getOption("googleAuthR.webapp.client_secret"))
}
\description{
Create a httr token from a js token
}
\keyword{internal}
