% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_mock.R
\name{gar_mock_list}
\alias{gar_mock_list}
\title{See which functions have mock cache}
\usage{
gar_mock_list()
}
\description{
When testing API calls, see which functions have API caches 
  activated when \code{options(googleAuthR.mock_test = TRUE)}
}
