% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopes.R
\name{gar_scope_config}
\alias{gar_scope_config}
\title{Create or add scopes to configuration}
\usage{
gar_scope_config(required_scopes)
}
\arguments{
\item{required_scopes}{character vector of scopes to add}
}
\description{
Helper for working with scopes
}
