% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard_auth.R
\name{gar_setup_auth_check}
\alias{gar_setup_auth_check}
\title{Check service key works via environment argument}
\usage{
gar_setup_auth_check(
  env_arg = "GCE_AUTH_FILE",
  scope = "https://www.googleapis.com/auth/cloud-platform"
)
}
\arguments{
\item{env_arg}{The authentication environment argument}

\item{scope}{The scope of the GCP request}
}
\description{
Check service key works via environment argument
}
\seealso{
Other setup functions: 
\code{\link{gar_setup_auth_key}()},
\code{\link{gar_setup_clientid}()},
\code{\link{gar_setup_edit_renviron}()},
\code{\link{gar_setup_env_check}()},
\code{\link{gar_setup_menu_do}()},
\code{\link{gar_setup_menu}()}
}
\concept{setup functions}
