% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_cache.R
\name{gar_setup_cache}
\alias{gar_setup_cache}
\title{Setup which package to perform mock testing upon}
\usage{
gar_setup_cache(package_name)
}
\arguments{
\item{package_name}{The package to cache API calls for

Only functions within the package will be mocked.  
The \code{options(googleAuthR.mock_test = TRUE)} flag will also need to be on.

You can also set the package via \code{options(googleAuthR.mock_package = "yourPackage")}}
}
\description{
This helps make sure the right functions are mocked
}
