% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard_auth.R
\name{gar_setup_clientid}
\alias{gar_setup_clientid}
\title{Check for a client JSON}
\usage{
gar_setup_clientid(session_user = NULL, client_json = "GAR_CLIENT_JSON")
}
\arguments{
\item{session_user}{1 for user level, 2 for project level, leave \code{NULL} to be prompted}

\item{client_json}{The environment argument to be used for client_id/secret}
}
\value{
TRUE is client_id is ready, FALSE if it is not
}
\description{
Check for a client JSON
}
\seealso{
Other setup functions: 
\code{\link{gar_setup_auth_check}()},
\code{\link{gar_setup_auth_key}()},
\code{\link{gar_setup_edit_renviron}()},
\code{\link{gar_setup_env_check}()},
\code{\link{gar_setup_menu_do}()},
\code{\link{gar_setup_menu}()}
}
\concept{setup functions}
