% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard.R
\name{gar_setup_edit_renviron}
\alias{gar_setup_edit_renviron}
\alias{gar_setup_check_session}
\title{Setup wizard help - asking users to edit .Renviron}
\usage{
gar_setup_edit_renviron(to_paste, session_user)

gar_setup_check_session(session_user = NULL)
}
\arguments{
\item{to_paste}{The line to paste into .Renviron}

\item{session_user}{whether its a 1 = user level or 2=project level .Renviron file

Intended to get user input from a menu, 1 indicating user level, 2 project level

\link{gar_setup_check_session} creates a menu for the user to choose which}
}
\description{
Setup wizard help - asking users to edit .Renviron
}
\examples{

\dontrun{

choice <- gar_setup_check_session()
gar_setup_edit_renviron("ENV_ARG=blah", session_user = choice)

}
}
\seealso{
Other setup functions: 
\code{\link{gar_setup_auth_check}()},
\code{\link{gar_setup_auth_key}()},
\code{\link{gar_setup_clientid}()},
\code{\link{gar_setup_env_check}()},
\code{\link{gar_setup_menu}()},
\code{\link{gar_setup_menu_do}()}
}
\concept{setup functions}
