% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard.R
\name{gar_setup_env_check}
\alias{gar_setup_env_check}
\title{Setup wizard help - check if environment argument is set}
\usage{
gar_setup_env_check(env_arg, set_to, edit_option = FALSE, session_user)
}
\arguments{
\item{env_arg}{The environment argument to check}

\item{set_to}{NULL or a string to set in .Renviron}

\item{edit_option}{Pass edit_option = FALSE to edit an existing environment arg}

\item{session_user}{1=user, 2=project scope of .Renviron}
}
\value{
TRUE once changes made
}
\description{
Setup wizard help - check if environment argument is set
}
\details{
Pass edit_option = FALSE to edit an existing environment arg, TRUE will 
  check if it exists, and will pass if its present.
}
\seealso{
Other setup functions: 
\code{\link{gar_setup_auth_check}()},
\code{\link{gar_setup_auth_key}()},
\code{\link{gar_setup_clientid}()},
\code{\link{gar_setup_edit_renviron}()},
\code{\link{gar_setup_menu}()},
\code{\link{gar_setup_menu_do}()}
}
\concept{setup functions}
