% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard_auth.R
\name{gar_setup_get_authenv}
\alias{gar_setup_get_authenv}
\title{Setup wizard helper - add authentication file to .Renviron}
\usage{
gar_setup_get_authenv(env_arg = "GCE_AUTH_FILE", ...)
}
\arguments{
\item{env_arg}{The environment argument to set}

\item{...}{Other arguments passed to \link{gar_setup_auth_key}}
}
\value{
A string to paste into an .Renviron, or NULL
}
\description{
Setup wizard helper - add authentication file to .Renviron
}
