% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_wizard.R
\name{gar_setup_menu}
\alias{gar_setup_menu}
\title{Setup wizard - introduction helper}
\usage{
gar_setup_menu(choices, package_name = "googleAuthR")
}
\arguments{
\item{choices}{A character vector of the choices passed to \link{menu}}

\item{package_name}{The package the setup menu is for}
}
\value{
The number option from the menu
}
\description{
Salutation and initial menu
}
\seealso{
Other setup functions: 
\code{\link{gar_setup_auth_check}()},
\code{\link{gar_setup_auth_key}()},
\code{\link{gar_setup_clientid}()},
\code{\link{gar_setup_edit_renviron}()},
\code{\link{gar_setup_env_check}()},
\code{\link{gar_setup_menu_do}()}
}
\concept{setup functions}
