% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_shiny_getUrl}
\alias{gar_shiny_getUrl}
\title{Get the Shiny Apps URL.}
\usage{
gar_shiny_getUrl(session)
}
\arguments{
\item{session}{The shiny session object.}
}
\value{
The URL of the Shiny App its called from.
}
\description{
Needed for the redirect URL in Google Auth flow
}
\seealso{
Other shiny auth functions: \code{\link{authReturnCode}};
  \code{\link{createCode}};
  \code{\link{gar_shiny_getAuthUrl}};
  \code{\link{gar_shiny_getToken}};
  \code{\link{loginOutput}};
  \code{\link{reactiveAccessToken}};
  \code{\link{renderLogin}}; \code{\link{with_shiny}}
}
\keyword{internal}

