% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{gar_shiny_login_ui}
\alias{gar_shiny_login_ui}
\title{A login page for Shiny}
\usage{
gar_shiny_login_ui(req, title = "googleAuthR Login Demo")
}
\arguments{
\item{req}{Used by shiny to generate login URL}

\item{title}{The title o}
}
\description{
An alternative to the immediate login provided by default by \link{gar_shiny_ui}
}
