% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gar_token_info}
\alias{gar_token_info}
\title{Get current token summary}
\usage{
gar_token_info(detail_level = getOption("googleAuthR.verbose", default =
  3))
}
\arguments{
\item{detail_level}{How much info to show}
}
\description{
Get details on the current active auth token to help debug issues
}
