% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{get_google_token_shiny}
\alias{get_google_token_shiny}
\title{Does the Shiny authentication flow}
\usage{
get_google_token_shiny(return_code, session = NULL)
}
\arguments{
\item{return_code}{the code in the url from \code{authReturnCode}}

\item{session}{A Shiny session object passed from shinyServer().}
}
\value{
A Google OAuth2 token
}
\description{
Used in a shiny session instead of \code{scr_auth}
}
\seealso{
Other shiny auth functions: \code{\link{authReturnCode}};
  \code{\link{createCode}};
  \code{\link{gar_shiny_getAuthUrl}};
  \code{\link{gar_shiny_getToken}};
  \code{\link{gar_shiny_getUrl}}
}

