% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{googleAuth_server_token}
\alias{googleAuth_server_token}
\title{googleAuth_server}
\usage{
googleAuth_server_token(session)
}
\arguments{
\item{session}{Shiny session argument}
}
\description{
If using googleAuth_ui, put this at the top of your server.R function
}
