% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{googleAuth_ui}
\alias{googleAuth_ui}
\title{googleAuth_ui}
\usage{
googleAuth_ui(req)
}
\arguments{
\item{req}{A Rook request}

\item{ui}{A Shiny ui object}
}
\description{
A function that will turn your ui object into one that will look for authentication
}
\details{
Put this at the bottom of your ui.R or pass into shinyApp()
}
