% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-auth.R
\name{googleAuthUI}
\alias{googleAuthUI}
\title{Shiny Google Authorisation [UI Module]}
\usage{
googleAuthUI(id)
}
\arguments{
\item{id}{shiny id}
}
\value{
A shiny UI for logging in
}
\description{
UI part of shiny module, use with \link{googleAuth}
}
\seealso{
Other shiny module functions: 
\code{\link{googleAuth}()}
}
\concept{shiny module functions}
