% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-js-signin.R
\name{googleSignInUI}
\alias{googleSignInUI}
\title{Google SignIn [UI Module]}
\usage{
googleSignInUI(id)
}
\arguments{
\item{id}{Shiny id}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{googleSignIn}. 
  If you just want a login to a Shiny app, without API tokens.
}
\seealso{
\url{https://github.com/dkulp2/Google-Sign-In}
}
\author{
Based on original code by David Kulp
}
