% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{gsc.is.valid.category.platform}
\alias{gsc.is.valid.category.platform}
\title{Checks valid parameters for Google Search Console}
\usage{
gsc.is.valid.category.platform(category, platform, include.all = FALSE)
}
\value{
TRUE if valid, raises an error if not.
}
\description{
The error type listings have certain categories/platform allowed.
}
\keyword{internal}

