% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{gsc.parseDimFilterGroup}
\alias{gsc.parseDimFilterGroup}
\title{Helper function for the query dimension filters}
\usage{
gsc.parseDimFilterGroup(dfe)
}
\arguments{
\item{dfe}{A string of form: dimension operator expression e.g. country!~GBR}
}
\description{
Saves the need for 3 parameters when you just have one
}
\details{
dimension = c('country','device','page','query')
  operator = c(`~~` = 'contains',
               `==` = 'equals',
               `!~` = 'notContains',
                `!=` = 'notEquals)

 expression = country: an ISO 3166-1 alpha-3 country code.
              device: 'DESKTOP','MOBILE','TABLET'
              page: not checked
              query: not checked
}
\seealso{
Other search analytics: \code{\link{lookupCountryCode}};
  \code{\link{search_analytics}}
}
\keyword{internal}

