% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\name{is_legit_token}
\alias{is_legit_token}
\title{Check that token appears to be legitimate}
\usage{
is_legit_token(x)
}
\description{
This unexported function exists to catch tokens that are technically valid,
i.e. `inherits(token, "Token2.0")` is TRUE, but that have dysfunctional
credentials.
}
\seealso{
Other authentication functions: \code{\link{gar_attach_auto_auth}},
  \code{\link{gar_auth_service}}, \code{\link{gar_auth}},
  \code{\link{gar_auto_auth}}, \code{\link{gar_gce_auth}},
  \code{\link{get_google_token}},
  \code{\link{token_exists}}
}
\keyword{internal}
