% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{is_shiny}
\alias{is_shiny}
\title{Is this a valid shiny session object?}
\usage{
is_shiny(shiny_session)
}
\arguments{
\item{shiny_session}{a Shiny session object.}
}
\value{
Boolean
}
\description{
Checks that a valid Shiny session object has been passed.
}
\keyword{internal}

