% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{list_sitemaps}
\alias{list_sitemaps}
\title{Gets sitemap information for the URL supplied.}
\usage{
list_sitemaps(siteUrl)
}
\arguments{
\item{siteURL}{The URL of the website to get sitemap information from. Must include protocol (http://).}

\item{shiny_access_token}{If in a Shiny session, supply its shiny_access_token object.}
}
\value{
A list of two dataframes: $sitemap with general info and $contents with sitemap info.
}
\description{
See here for details: https://developers.google.com/webmaster-tools/v3/sitemaps
}
\details{
In a shiny session, wrap the function in \code{with_shiny} and supply the session object
  in a \code{shiny_access_token} parameter
}
\seealso{
Other data fetching functions: \code{\link{checkTokenAPI}};
  \code{\link{delete_website}}
}

