% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{list_websites}
\alias{list_websites}
\title{Retrieves dataframe of websites user has in Search Console}
\usage{
list_websites(path_arguments = NULL, pars_arguments = NULL,
  shiny_access_token = NULL, the_body = NULL, ...)
}
\arguments{
\item{shiny_access_token}{If in a Shiny session, supply its shiny_access_token object.}
}
\value{
a dataframe of siteUrl and permissionLevel
}
\description{
Retrieves dataframe of websites user has in Search Console
}
\seealso{
Other search console website functions: \code{\link{add_website}};
  \code{\link{delete_website}}
}

