% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{loginOutput}
\alias{loginOutput}
\title{Login/logout Shiny output}
\usage{
loginOutput(output_name)
}
\arguments{
\item{output_name}{Name of what output object was assigned in \code{renderLogin}}
}
\value{
A login/logout button in a Shiny app
}
\description{
USe within a ui.R to render the login button generated by \code{renderLogin}
}
\examples{
\dontrun{
## in server.R
shinyServer(function(input, output, session)){

  ## Get auth code from return URL
  access_token  <- reactiveAccessToken(session)

  ## Make a loginButton to display using loginOutput
  output$loginButton <- renderLogin(session, access_token())

  output$websites <- renderTable({

    ## using with_shiny adds a shiny_access_token parameter to pass access_token()
    list_websites <- with_shiny(list_websites,
                                access_token())

    })

}

## in ui.R
shinyUI(fluidPage(
  loginOutput("loginButton"),
  renderTable("websites)
  ))
}
}
\seealso{
Other shiny auth functions: \code{\link{authReturnCode}};
  \code{\link{createCode}};
  \code{\link{gar_shiny_getAuthUrl}};
  \code{\link{gar_shiny_getToken}};
  \code{\link{gar_shiny_getUrl}};
  \code{\link{reactiveAccessToken}};
  \code{\link{renderLogin}}; \code{\link{with_shiny}}
}

