% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{lookupCountryCode}
\alias{lookupCountryCode}
\title{Look up the country codes}
\usage{
lookupCountryCode(country.code, name_type = "Name")
}
\arguments{
\item{name_type}{One of "Name", "Official_name" or "Common_name"}

\item{code}{the ISO 3166-1 alpha-3 country code (as used in searchConsoleR)}
}
\value{
The country name string
}
\description{
Look up the country codes
}
\seealso{
Other search analytics: \code{\link{gsc.parseDimFilterGroup}};
  \code{\link{search_analytics}}
}
\keyword{internal}

