% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{makeBatchRequest}
\alias{makeBatchRequest}
\title{Make the batch request inner content}
\usage{
makeBatchRequest(f)
}
\arguments{
\item{f}{The original unbatched Google API function call generated by \code{gar_api_generator}}
}
\description{
Make the batch request inner content
}
\seealso{
Other batch functions: \code{\link{applyDataParseFunction}},
  \code{\link{doBatchRequest}},
  \code{\link{gar_batch_walk}}, \code{\link{gar_batch}},
  \code{\link{parseBatchResponse}}
}
\keyword{internal}

