% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_searchConsole.R
\name{parse_search_analytics}
\alias{parse_search_analytics}
\title{Parsing search_analytics data}
\usage{
parse_search_analytics(x, dimensions, prettyNames)
}
\arguments{
\item{x}{the req$content response}

\item{dimensions}{the dimensions}

\item{prettyNames}{whether to translate the country codes}
}
\description{
Parsing search_analytics data
}

