% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{parseBatchResponse}
\alias{parseBatchResponse}
\title{Parse batch request}
\usage{
parseBatchResponse(batch_response)
}
\arguments{
\item{batch_response}{An element of the list of responses from a batched request}
}
\description{
Parse batch request
}
\seealso{
Other batch functions: \code{\link{applyDataParseFunction}},
  \code{\link{doBatchRequest}},
  \code{\link{gar_batch_walk}}, \code{\link{gar_batch}},
  \code{\link{makeBatchRequest}}
}
\keyword{internal}
