% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{postwalk}
\alias{postwalk}
\title{Walk into a list}
\usage{
postwalk(x, func)
}
\arguments{
\item{x}{what to check}

\item{func}{Function to apply if not a list}
}
\value{
the function acting on x or an inner element of x
}
\description{
If passed an object such as a nested list, will apply function
  on inner elements that are not lists.
}
\keyword{internal}

