% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{replace.kv}
\alias{replace.kv}
\title{Create a modified list}
\usage{
replace.kv(template, replace)
}
\arguments{
\item{template}{a (nested) list with elements to replace}

\item{replace}{a subset of template with same names but replacement values}
}
\value{
a list like template but with values replace from replace
}
\description{
Create a modified list
}
\keyword{internal}

