% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{retryRequest}
\alias{retryRequest}
\title{ReTry API requests for certain errors using exponential backoff.}
\usage{
retryRequest(f)
}
\arguments{
\item{f}{A function of a http request}
}
\description{
ReTry API requests for certain errors using exponential backoff.
}
\keyword{internal}
