% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{revokeEventObserver}
\alias{revokeEventObserver}
\title{Listens for a user revoking authentication}
\usage{
revokeEventObserver(access_token, input)
}
\arguments{
\item{access_token}{A token generated by \code{reactiveAccessToken}.}

\item{input}{the input object from a shinyServer function.}
}
\description{
If the parameter \code{revoke} is set to TRUE for \link{renderLogin}
  then this observer is also required in the Shiny server to do the revoking.
}
\seealso{
\link{renderLogin}

Other shiny auth functions: \code{\link{authReturnCode}},
  \code{\link{createCode}},
  \code{\link{gar_shiny_getAuthUrl}},
  \code{\link{gar_shiny_getToken}},
  \code{\link{gar_shiny_getUrl}},
  \code{\link{loginOutput}},
  \code{\link{reactiveAccessToken}},
  \code{\link{renderLogin}}, \code{\link{with_shiny}}
}

