% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\name{scr_auth_suspend}
\alias{scr_auth_suspend}
\title{Suspend authorization}
\usage{
scr_auth_suspend(disable_httr_oauth = TRUE, verbose = TRUE)
}
\description{
This unexported function exists so we can suspend authorization for
testing purposes.
}
\keyword{internal}

