% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_urlshort.R
\name{shorten_url}
\alias{shorten_url}
\title{Shortens a url using goo.gl}
\usage{
shorten_url(url)
}
\arguments{
\item{url}{URl to shorten with goo.gl}
}
\value{
a string of the short URL
}
\description{
Shortens a url using goo.gl
}

