% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{silent_auth}
\alias{silent_auth}
\title{Silent auth}
\usage{
silent_auth(req)
}
\arguments{
\item{req}{What Shiny uses to check the URL parameters}
}
\description{
The default for logging in via \link{gar_shiny_ui}, this creates no login page and just takes you straight to authentication on Shiny app load.
}
\seealso{
Other pre-load shiny authentication: \code{\link{gar_shiny_auth_url}},
  \code{\link{gar_shiny_auth}},
  \code{\link{gar_shiny_login_ui}},
  \code{\link{gar_shiny_ui}}
}
