% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{skip_if_no_env_auth}
\alias{skip_if_no_env_auth}
\title{Skip test if not authenticated}
\usage{
skip_if_no_env_auth(env_arg)
}
\arguments{
\item{env_arg}{The name of the environment argument pointing to the auth file}
}
\description{
Use within tests to skip if a local authentication file isn't available through an environment variable.
}
