% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{split_vector}
\alias{split_vector}
\title{split a vector on an element}
\usage{
split_vector(vector, index, remove_splits = TRUE)
}
\arguments{
\item{vector}{a vector}

\item{index}{where to split the vector}

\item{remove_splits}{to return with or without the index points}
}
\value{
a list of vectors split not including split point
}
\description{
Why doesn't this exist?
}
\keyword{internal}
