% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{substitute.list}
\alias{substitute.list}
\title{Substitute in a (nested) list}
\usage{
substitute.list(template, replace_me)
}
\arguments{
\item{template}{A template named list}

\item{replace_me}{A similar named list with different values to substitute}
}
\value{
The template with the values substituted.

If replace_me has list names not in template, the value stays the same.
}
\description{
Substitute in a (nested) list
}

