% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{token_exists}
\alias{token_exists}
\title{Check if authorization currently in force}
\usage{
token_exists()
}
\value{
logical
}
\description{
Check if authorization currently in force
}
\seealso{
Other authentication functions: \code{\link{gar_attach_auto_auth}},
  \code{\link{gar_auth_service}}, \code{\link{gar_auth}},
  \code{\link{gar_auto_auth}}, \code{\link{gar_gce_auth}},
  \code{\link{get_google_token}}
}
\keyword{internal}
