% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_urlshort.R
\name{user_history}
\alias{user_history}
\title{Get the history of the authenticated user}
\usage{
user_history()
}
\value{
a dataframe of the goo.gl user's history
}
\description{
Get the history of the authenticated user
}

