% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{applyDataParseFunction}
\alias{applyDataParseFunction}
\title{Apply parsing function if a good response}
\usage{
applyDataParseFunction(function_entry, batch_content, ...)
}
\arguments{
\item{function_entry}{a Google API function generated by \code{gar_api_generator}}

\item{the}{content returned from a batch call}
}
\description{
Apply parsing function if a good response
}
\seealso{
Other batch functions: \code{\link{doBatchRequest}},
  \code{\link{gar_batch_walk}}, \code{\link{gar_batch}},
  \code{\link{makeBatchRequest}},
  \code{\link{parseBatchResponse}}
}
\keyword{internal}
