% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\docType{data}
\name{Authentication}
\alias{Authentication}
\title{R6 environment to store authentication credentials}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x10dbc5158> 
 - attr(*, "name")= chr "Authentication_generator"
}}
\usage{
Authentication
}
\description{
Used to keep persistent state.
}
\keyword{datasets}

