% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\docType{data}
\name{Authentication}
\alias{Authentication}
\title{R6 environment to store authentication credentials}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Authentication
}
\description{
Used to keep persistent state.
}
\keyword{datasets}
