% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checkTokenAPI}
\alias{checkTokenAPI}
\title{Check API data token}
\usage{
checkTokenAPI()
}
\value{
boolean if it works.
}
\description{
Check API data token
}
\concept{data fetching functions}
\keyword{internal}
