% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{doBatchRequest}
\alias{doBatchRequest}
\title{Batch Requests}
\usage{
doBatchRequest(batched)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{session}{a shiny session object if in a shiny app.}

\item{request_type}{the type of httr request function: GET, POST, PUT, DELETE etc.}

\item{the_body}{body of POST request}

\item{params}{A named character vector of other parameters to add to request.}
}
\description{
Batch requests to Google APIs that support it
}
\details{
Example of params: c(param1="foo", param2="bar")
}
\keyword{internal}

