% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{doBatchRequest}
\alias{doBatchRequest}
\title{Batch Requests}
\usage{
doBatchRequest(batched)
}
\arguments{
\item{batched}{an element of a list of parsed batch requests}
}
\description{
Batch requests to Google APIs that support it
}
\seealso{
Other batch functions: \code{\link{applyDataParseFunction}},
  \code{\link{gar_batch_walk}}, \code{\link{gar_batch}},
  \code{\link{makeBatchRequest}},
  \code{\link{parseBatchResponse}}
}
\keyword{internal}

