% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{gar_auth}
\alias{gar_auth}
\title{Authorize \code{googleAuthR}}
\usage{
gar_auth(token = NULL, email = NULL,
  scopes = getOption("googleAuthR.scopes.selected"), new_user = NULL,
  package = "googleAuthR")
}
\arguments{
\item{token}{an actual token object or the path to a valid token stored as an
\code{.rds} file}

\item{email}{An existing gargle cached email to authenticate with or TRUE to authenticate with the only email available.}

\item{scopes}{Scope of the request}

\item{new_user}{Not used}

\item{package}{The name of the package authenticating}
}
\value{
an OAuth token object, specifically a
  \code{\link[=Token-class]{Token2.0}}, invisibly
}
\description{
Wrapper of \link[gargle]{token_fetch}
}
\seealso{
Other authentication functions: \code{\link{gar_attach_auto_auth}},
  \code{\link{gar_auth_service}},
  \code{\link{gar_auto_auth}}, \code{\link{gar_gce_auth}},
  \code{\link{get_google_token}},
  \code{\link{token_exists}}
}
\concept{authentication functions}
