% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_auth_js}
\alias{gar_auth_js}
\title{Shiny JavaScript Google Authorisation [Server Module]}
\usage{
gar_auth_js(input, output, session)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}
}
\value{
A httr reactive OAuth2.0 token
}
\description{
Shiny Module for use with \link{gar_auth_jsUI}
}
\details{
Call via \code{shiny::callModule(gar_auth_js, "your_id")}
}
