% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_shiny.R
\name{gar_auth_jsUI}
\alias{gar_auth_jsUI}
\title{Shiny JavaScript Google Authorisation [UI Module]}
\usage{
gar_auth_jsUI(id, login_class = "btn btn-primary",
  logout_class = "btn btn-danger", login_text = "Log In",
  logout_text = "Log Out")
}
\arguments{
\item{id}{Shiny id}

\item{login_class}{CSS class of login button}

\item{logout_class}{CSS class ofr logout button}

\item{login_text}{Text to show on login button}

\item{logout_text}{Text to show on logout button}
}
\value{
Shiny UI
}
\description{
A Javascript Google authorisation flow for Shiny apps.
}
\details{
Shiny Module for use with \link{gar_auth_js}
}

