% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\name{gar_auth_service}
\alias{gar_auth_service}
\title{Create a service account authentication}
\usage{
gar_auth_service(secrets)
}
\arguments{
\item{secrets}{the JSON file downloaded from Google search console}
}
\value{
(Invisible) Sets authentication token
}
\description{
Create a service account authentication
}
\seealso{
\code{\link{https://developers.google.com/identity/protocols/OAuth2ServiceAccount}}
}

