% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{gar_batch}
\alias{gar_batch}
\title{Turn a list of gar_fetch_functions into batch functions}
\usage{
gar_batch(function_list)
}
\arguments{
\item{a}{list of functions generated by \code{gar_api_generator}}
}
\value{
A list of the Google API responses
}
\description{
Some Google APIs support batch processing, which dramatically speeds up response
  when you have multiple calls to make.  This function allows you to take advatage of this.
}
\details{
This function will turn all the individual Google API functions
  into one POST request to /batch.
}
\examples{
\dontrun{
# from search console API
list_websites <- function() {
  l <- gar_api_generator("https://www.googleapis.com/webmasters/v3/sites",
                         "GET",
                         data_parse_function = function(x) x$siteEntry)
  l()
  }

 from goo.gl API
 user_history <- function(){
   f <- gar_api_generator("https://www.googleapis.com/urlshortener/v1/url/history",
                          "GET",
                          data_parse_function = function(x) x$items)

  f()
  }

gar_auth()
ggg <- gar_batch(list(list_websites(), user_history())))
ggg


}
}
\seealso{
\code{\link{https://developers.google.com/webmaster-tools/v3/how-tos/batch}}
  Documentation on doing batch requests for the search console API.
  Other Google APIs are similar.

Other batch functions: \code{\link{applyDataParseFunction}};
  \code{\link{doBatchRequest}};
  \code{\link{makeBatchRequest}};
  \code{\link{parseBatchResponse}}
}

