% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_batch.R
\name{gar_batch}
\alias{gar_batch}
\title{Turn a list of gar_fetch_functions into batch functions}
\usage{
gar_batch(call_list, ...)
}
\arguments{
\item{call_list}{a list of functions from \code{\link{gar_api_generator}}}

\item{...}{further arguments passed to the data parse function of f}
}
\value{
A list of the Google API responses
}
\description{
Turn a list of gar_fetch_functions into batch functions
}
\details{
This function will turn all the individual Google API functions
  into one POST request to /batch. 
  
If you need to pass multiple data parse function arguments its probably best 
  to do it in seperate batches to avoid confusion.
}
\seealso{
https://developers.google.com/webmaster-tools/v3/how-tos/batch

  Documentation on doing batch requests for the search console API.  
  Other Google APIs are similar.
  
  Walk through API calls changing parameters using \code{\link{gar_batch_walk}}

Other batch functions: \code{\link{applyDataParseFunction}},
  \code{\link{doBatchRequest}},
  \code{\link{gar_batch_walk}},
  \code{\link{makeBatchRequest}},
  \code{\link{parseBatchResponse}}
}
