% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_memoise.R
\name{gar_cache_setup}
\alias{gar_cache_setup}
\title{Setup which package to perform mock testing upon}
\usage{
gar_cache_setup(mcache = memoise::cache_memory())
}
\arguments{
\item{mcache}{A cache method from \link[memoise]{memoise}. If \code{NULL} will empty.}
}
\description{
If set, googleAuthR dependecies will start caching API responses
}
\details{
This helps make sure the right functions are mocked
}
