% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_cache.R
\name{gar_cache_setup}
\alias{gar_cache_setup}
\title{Setup which package to perform mock testing upon}
\usage{
gar_cache_setup(package_name, location = "memory")
}
\arguments{
\item{package_name}{The package to cache API calls for}

\item{location}{Where to put the cache.  Use \code{memory} to place in RAM, other location of directory

Only functions within the package will be mocked.  

You can also set the package via \code{options(googleAuthR.mock_package = "yourPackage")}}
}
\description{
This helps make sure the right functions are mocked
}
