% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_memoise.R
\name{gar_cache_setup}
\alias{gar_cache_setup}
\title{Setup where to put cache}
\usage{
gar_cache_setup(mcache = memoise::cache_memory())
}
\arguments{
\item{mcache}{A cache method from \link[memoise]{memoise}. If \code{NULL} will do no caching.}
}
\description{
To cache to a file system use \code{memoise::cache_filesystem("cache_folder")}, 
  suitable for unit testing and works between R sessions. 

The cached API calls do not need authentication to be active, but need this function to set caching first.
}
