% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{gar_create_api_objects}
\alias{gar_create_api_objects}
\title{Create the API objects from the Discovery API}
\usage{
gar_create_api_objects(filename, api_json, format = TRUE)
}
\arguments{
\item{filename}{File to write the objects to}

\item{api_json}{The json from \link{gar_discovery_api}}

\item{format}{If TRUE will use \link[formatR]{tidy_eval} on content}
}
\value{
TRUE if successful, side-effect creating filename
}
\description{
Create the API objects from the Discovery API
}
\seealso{
Other Google Discovery API functions: 
\code{\link{gar_create_api_skeleton}()},
\code{\link{gar_create_package}()},
\code{\link{gar_discovery_apis_list}()},
\code{\link{gar_discovery_api}()}
}
\concept{Google Discovery API functions}
