% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{gar_create_api_skeleton}
\alias{gar_create_api_skeleton}
\title{Create an API library skeleton}
\usage{
gar_create_api_skeleton(filename = "./R/new_api.R",
  api_json = gar_discovery_api("urlshortener", "v1"),
  api_method = api_json$resources$url$methods)
}
\arguments{
\item{filename}{R file to write skeleton to}

\item{api_json}{The json from \link{gar_discovery_api}}

\item{api_method}{Location of API resource methods}
}
\value{
TRUE if successful, side effect will write a file
}
\description{
This will create a file with the skeleton of the API functions 
  for the specified library
}

