% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_discovery.R
\name{gar_create_api_skeleton}
\alias{gar_create_api_skeleton}
\title{Create an API library skeleton}
\usage{
gar_create_api_skeleton(filename, api_json, format = TRUE)
}
\arguments{
\item{filename}{R file to write skeleton to}

\item{api_json}{The json from \link{gar_discovery_api}}

\item{format}{If TRUE will use \link[formatR]{tidy_eval} on content}
}
\value{
TRUE if successful, side effect will write a file
}
\description{
This will create a file with the skeleton of the API functions 
  for the specified library
}
\seealso{
Other Google Discovery API functions: \code{\link{gar_create_api_objects}},
  \code{\link{gar_create_package}},
  \code{\link{gar_discovery_apis_list}},
  \code{\link{gar_discovery_api}}
}

