% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{gar_discovery_api}
\alias{gar_discovery_api}
\title{Get meta data details for specified Google API}
\usage{
gar_discovery_api(api, version)
}
\arguments{
\item{api}{The API to fetch}

\item{version}{The API version to fetch}
}
\value{
Details of the API
}
\description{
Does not require authentication
}
\seealso{
\url{https://developers.google.com/discovery/v1/reference/apis/getRest}

Other Google Discovery API functions: \code{\link{gar_create_api_objects}},
  \code{\link{gar_create_api_skeleton}},
  \code{\link{gar_create_package}},
  \code{\link{gar_discovery_apis_list}}
}
