% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{gar_discovery_api}
\alias{gar_discovery_api}
\title{Get meta data details for specified Google API}
\usage{
gar_discovery_api(api, version, a_url = NULL)
}
\arguments{
\item{api}{The API to fetch}

\item{version}{The API version to fetch}

\item{a_url}{Supply your own discovery URL, for private APIs only}
}
\value{
Details of the API
}
\description{
Download the discovery document for an API
}
\seealso{
\url{https://developers.google.com/discovery/v1/reference/apis/getRest}

Other Google Discovery API functions: 
\code{\link{gar_create_api_objects}()},
\code{\link{gar_create_api_skeleton}()},
\code{\link{gar_create_package}()},
\code{\link{gar_discovery_apis_list}()}
}
\concept{Google Discovery API functions}
