% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_discovery.R
\name{gar_discovery_apis_list}
\alias{gar_discovery_apis_list}
\title{Get a list of Google API libraries}
\usage{
gar_discovery_apis_list()
}
\value{
List of Google APIs and their resources
}
\description{
Doesn't require authentication
}
\seealso{
\url{https://developers.google.com/discovery/v1/reference/apis/list}

Other Google Discovery API functions: \code{\link{gar_create_api_objects}},
  \code{\link{gar_create_api_skeleton}},
  \code{\link{gar_create_package}},
  \code{\link{gar_discovery_api}}
}

