% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-modifyurl.R
\name{gar_shiny_login_ui}
\alias{gar_shiny_login_ui}
\title{A login page for Shiny}
\usage{
gar_shiny_login_ui(req, title = "googleAuthR Login Demo")
}
\arguments{
\item{req}{Passed to \link{gar_shiny_auth_url} to generate login URL}

\item{title}{The title of the page}
}
\description{
An alternative to the immediate login provided by default by \link{gar_shiny_ui}
}
\details{
Use \link{gar_shiny_auth_url} to create the login URL.  You must leave the first argument free as this is used to generate the login, but you can pass other arguments to customise your UI.
}
\seealso{
Other pre-load shiny authentication: 
\code{\link{gar_shiny_auth}()},
\code{\link{gar_shiny_auth_url}()},
\code{\link{gar_shiny_ui}()},
\code{\link{silent_auth}()}
}
\concept{pre-load shiny authentication}
