% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_google_token}
\alias{get_google_token}
\title{Retrieve Google token from environment and configs for httr}
\usage{
get_google_token()
}
\value{
a httr configured option for token
For shiny the token is passed from reactive session
}
\description{
Get token if it's previously stored, else prompt user to get one.
}
\seealso{
Other authentication functions: 
\code{\link{gar_attach_auto_auth}()},
\code{\link{gar_auth}()},
\code{\link{gar_auth_service}()},
\code{\link{gar_auto_auth}()},
\code{\link{gar_gce_auth}()},
\code{\link{should_skip_token_checks}()},
\code{\link{token_exists}()}
}
\concept{authentication functions}
\keyword{internal}
