% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_auth.R
\name{get_google_token}
\alias{get_google_token}
\title{Retrieve Google token from environment and configs for httr}
\usage{
get_google_token(shiny_return_token = NULL)
}
\arguments{
\item{shiny_return_token}{In a shiny session, this is passed instead.}
}
\value{
a httr configured option for token
For shiny the token is passed from reactive session
}
\description{
Get token if it's previously stored, else prompt user to get one.
}
\seealso{
Other authentication functions: \code{\link{gar_auth_service}},
  \code{\link{gar_auth}}, \code{\link{is_legit_token}},
  \code{\link{token_exists}}
}
\keyword{internal}

