% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-js-auth.R
\name{googleAuth_js}
\alias{googleAuth_js}
\title{Shiny JavaScript Google Authorisation [Server Module]}
\usage{
googleAuth_js(
  input,
  output,
  session,
  message = "Authenticate with your Google account"
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{message}{The message to show when not authenticated}
}
\value{
A httr reactive OAuth2.0 token
}
\description{
Shiny Module for use with \link{googleAuth_jsUI}
}
\details{
Call via \code{shiny::callModule(googleAuth_js, "your_id")}
}
