% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-js-signin.R
\name{googleSignIn}
\alias{googleSignIn}
\title{Google SignIn [Server Module]}
\usage{
googleSignIn(input, output, session)
}
\arguments{
\item{input}{shiny input (must contain \code{g_id}, \code{g_name},
\code{g_email}, \code{g_image}, \code{g_signed_in})}

\item{output}{shiny output (passed by shiny but not used)}

\item{session}{shiny session}
}
\value{
A reactive list with values \code{$id}, \code{$name}, \code{$email},
  \code{$image} and \code{$signed_in}.
}
\description{
Shiny Module for use with \link{googleSignInUI}.  
  Use when you don't need to call APIs, but would like a login to Shiny.
}
\details{
Call via \code{shiny::callModule(googleSignIn, "your_id")}.
}
\author{
Based on original code by David Kulp
}
