% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-js-signin.R
\name{googleSignInUI}
\alias{googleSignInUI}
\title{Google SignIn [UI Module]}
\usage{
googleSignInUI(id, logout_name = "Sign Out",
  logout_class = "btn-danger")
}
\arguments{
\item{id}{Shiny id.}

\item{logout_name}{Character. Custom name of the logout button.}

\item{logout_class}{Character. Bootstrap class name for buttons, e.g. "btn-info", "btn-dark".}
}
\value{
Shiny UI
}
\description{
Shiny Module for use with \link{googleSignIn}. 
  If you just want a login to a Shiny app, without API tokens.
}
\seealso{
\url{https://github.com/dkulp2/Google-Sign-In}
}
\author{
Based on original code by David Kulp
}
