% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleAuthR_utility.R
\name{substitute.list}
\alias{substitute.list}
\title{Substitute in a (nested) list}
\usage{
substitute.list(template, replace_me)
}
\arguments{
\item{template}{A template named list}

\item{replace_me}{A similar named list with different values to substitute}
}
\value{
The template with the values substituted.
}
\description{
Substitute in a (nested) list
}
\keyword{If}
\keyword{has}
\keyword{in}
\keyword{internal}
\keyword{list}
\keyword{names}
\keyword{not}
\keyword{replace_me}
\keyword{same.}
\keyword{stays}
\keyword{template,}
\keyword{the}
\keyword{value}
